offlinemap/requirements_clean.txt
bag-convert==0.0.1
bottleneck
catkin-pkg==1.0.0
certifi
cffi==1.17.1
contourpy==1.3.2
cyber-record==0.1.12
cycler==0.12.1
distro==1.9.0
docutils==0.21.2
et_xmlfile==2.0.0
fonttools==4.58.4
geopandas==1.1.1
gnupg==2.3.1
greenlet
kiwisolver==1.4.8
lzf==0.1
matplotlib==3.10.3
mkl-service
mkl_fft
mkl_random
networkx
numexpr
numpy>=2.0
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas
pillow==11.2.1
protobuf==3.19.4
psutil==7.0.0
psycopg2-binary
py3rosmsgs==1.18.2
pycparser==2.22
pycryptodomex==3.23.0
pyogrio==0.11.0
pyparsing==3.2.3
pyproj
python-dateutil
pytz
PyYAML==6.0.2
record-msg==0.1.1
rospkg==1.6.0
setuptools==78.1.1
shapely
six
SQLAlchemy
typing_extensions
tzdata
wheel==0.45.1